Complete backend and put work into front-end.
So far, I like how the project is coming along. I like the color
scheme and the look of everything. I first drew up what I thought
the UI should look like and tried to stay true to my vision.
Had to change up the dates on the ambitions as there wasn't enough
width to fit them in the way I was envisioning.

I don't think anything needs to be changed (except for icons instead
of singular characters), but everything just looks good so far.
Just need to work out the generation of ambitions, goals, tasks,
and subTasks. I've learned a lot already with this project and I'm
trying to stick with the SOLID principles and modularity.

Needs to be worked on/finished:
    Storage backup and loading
    Automatic generation of ambition tiles based on array
    Ability to edit everything
    Clicking on status bar brings up the in depth views.
    Icons for Completion, Editing, and Deletion