I don't really know what happened... It feels like I went back in time. Every thing is old? or not how I thought it was supposed to be at least.
Well, it's fixed? I think? I don't understand either because the project
was working the last time I left it. It's still working online? And the
code looks different from how it looks now?

I don't know if a commit got reverted at some point, but it seems like
a lot of progress is deleted (was?)